+++ 001_schema_migrations.down.sql
@@ -1,0 +1,2 @@
+-- Reverses 001_schema_migrations.sql
+DROP TABLE IF EXISTS schema_migrations;